+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 All notable changes to `Browsershot` will be documented in this file
 
+## 3.42.0 - 2021-01-11
+
+- introduce a selectorIndex to bypass querySelector restrictions (#468)
+
+## 3.41.2 - 2020-12-27
+
+- improve local require for puppeteer (#461)
+
+## 3.41.1 - 2020-12-08
+
+- replace `waitFor` with `waitForTimeout` (#452)
+
 ## 3.41.0 - 2020-19-11
 
 - adding ability to pass envars to browser instance (#448)